+++ packages/sdk/src/swap/sdk.ts
@@ -175,7 +175,7 @@ export class SwapSDK {
       '[@chainflip/sdk] The getQuote method is deprecated and will be removed in the 1.9 release. Use the getQuoteV2 method instead.',
     );
 
-    const { brokerCommissionBps, affiliateBrokers, ...remainingRequest } = quoteRequest;
+    const { brokerCommissionBps, affiliateBrokers, ccmParams, ...remainingRequest } = quoteRequest;
     const submitterBrokerCommissionBps =
       brokerCommissionBps ?? this.options.broker?.commissionBps ?? 0;
     const affiliateBrokerCommissionBps =
@@ -185,6 +185,8 @@ export class SwapSDK {
       this.options.backendUrl,
       {
         ...remainingRequest,
+        ccmGasBudget: ccmParams?.gasBudget,
+        ccmMessageLengthBytes: ccmParams?.messageLengthBytes,
         brokerCommissionBps: submitterBrokerCommissionBps + affiliateBrokerCommissionBps,
       },
       options,
@@ -195,7 +197,7 @@ export class SwapSDK {
     quoteRequest: QuoteRequest,
     options: ApiService.RequestOptions = {},
   ): Promise<QuoteResponseV2> {
-    const { brokerCommissionBps, affiliateBrokers, ...remainingRequest } = quoteRequest;
+    const { brokerCommissionBps, affiliateBrokers, ccmParams, ...remainingRequest } = quoteRequest;
     const submitterBrokerCommissionBps =
       brokerCommissionBps ?? this.options.broker?.commissionBps ?? 0;
     const affiliateBrokerCommissionBps =
@@ -206,6 +208,8 @@ export class SwapSDK {
       {
         ...remainingRequest,
         brokerCommissionBps: submitterBrokerCommissionBps + affiliateBrokerCommissionBps,
+        ccmGasBudget: ccmParams?.gasBudget,
+        ccmMessageLengthBytes: ccmParams?.messageLengthBytes,
         dcaEnabled: this.dcaEnabled,
       },
       options,
@@ -402,6 +406,12 @@ export class SwapSDK {
     assertQuoteValid(quote);
     assert(!quote.isVaultSwap, 'Cannot open a deposit channel for a vault swap quote');
 
+    if (ccmParams) {
+      assert(quote.ccmParams, 'Cannot open CCM channel for quote without CCM params');
+    } else {
+      assert(!quote.ccmParams, 'Cannot open regular channel for quote with CCM params');
+    }
+
     const depositAddressRequest = {
       srcAsset: quote.srcAsset.asset,
       srcChain: quote.srcAsset.chain,
@@ -480,6 +490,12 @@ export class SwapSDK {
     await this.validateSwapAmount(quote.srcAsset, BigInt(quote.depositAmount));
     assertQuoteValid(quote);
     assert(quote.isVaultSwap, 'Cannot encode vault swap data for a deposit channel quote');
+
+    if (ccmParams) {
+      assert(quote.ccmParams, 'Cannot encode CCM swap for quote without CCM params');
+    } else {
+      assert(!quote.ccmParams, 'Cannot encode regular swap for quote with CCM params');
+    }
 
     const vaultSwapRequest = {
       srcAsset: quote.srcAsset,