+++ packages/swap/src/routes/__tests__/quote.test.ts
@@ -494,7 +494,7 @@ describe('server', () => {
         hexEncodeNumber(100_000_000), // deposit amount
         10,
         undefined,
-        null,
+        undefined,
         [],
         [],
       );
@@ -571,7 +571,7 @@ describe('server', () => {
         hexEncodeNumber(100_000_000), // deposit amount
         0,
         undefined,
-        null,
+        undefined,
         ['IngressDepositChannel'],
         [],
       );
@@ -673,7 +673,7 @@ describe('server', () => {
         hexEncodeNumber(1e8), // deposit amount
         0,
         undefined,
-        null,
+        undefined,
         [],
         [],
       );
@@ -686,7 +686,7 @@ describe('server', () => {
         hexEncodeNumber(99950000), // deposit amount - boost fee
         0,
         undefined,
-        null,
+        undefined,
         [],
         [],
       );
@@ -741,7 +741,7 @@ describe('server', () => {
         hexEncodeNumber(1e8), // deposit amount
         0,
         undefined,
-        null,
+        undefined,
         [],
         [],
       );
@@ -820,7 +820,7 @@ describe('server', () => {
         hexEncodeNumber(100e6), // deposit amount,
         0,
         undefined,
-        null,
+        undefined,
         [],
         [],
       );